// src/modules/notification/enums/notification-event.enum.ts
import { registerEnumType } from '@nestjs/graphql';

export enum NotificationEvent {
    ASSESSMENT_ASSIGNED = 'ASSESSMENT_ASSIGNED',
    ASSESSMENT_REMINDER = 'ASSESSMENT_REMINDER',
    ASSESSMENT_ANSWERED = 'ASSESSMENT_ANSWERED',
    ASSESSMENT_NOT_ANSWERED = 'ASSESSMENT_NOT_ANSWERED',
    ASSESSMENT_PERIODIC_SUMMARY = 'ASSESSMENT_PERIODIC_SUMMARY',
    CASE_UPDATED = 'CASE_UPDATED',
    USER_CREATED = 'USER_CREATED',
    FIRST_LOGIN = 'FIRST_LOGIN',
    LAST_LOGIN = 'LAST_LOGIN',
    SESSION_NUMBER = 'SESSION_NUMBER',
    TREATMENT_FINALIZATION = 'TREATMENT_FINALIZATION',
    SESSION_NO_SHOW_CANCELLATION = 'SESSION_NO_SHOW_CANCELLATION',
    NEW_TREATMENT = 'NEW_TREATMENT',
}

registerEnumType(NotificationEvent, {
    name: 'NotificationEvent',
});